Network code clean-up.
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@5896 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/core/org/jnode/vm/classmgr/SelectorMap.java
@@ -46,7 +46,7 @@ public class SelectorMap extends VmSystemObject {
         final String id = InternString.internString(name + '#' + signature);
         final Integer selector = (Integer) map.get(id);
         if (selector != null) {
-            return selector.intValue();
+            return selector;
         } else {
             return getNew(id);
         }
@@ -62,7 +62,7 @@ public class SelectorMap extends VmSystemObject {
     private synchronized int getNew(String id) {
         Integer selector = (Integer) map.get(id);
         if (selector != null) {
-            return selector.intValue();
+            return selector;
         } else {
             final int sel = ++lastSelector;
             map.put(id, sel);
